# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "新能源材料进销存 API",
    "description": "进销存后端接口：用户与库房、收货定价、对外报价、入库审核、出库与 FIFO、库存报表与预警等。默认账号见服务启动日志。详细说明见仓库内 `server/API接口文档.md`。\n\n**鉴权**：除「健康检查」「登录」「自助注册」外，请求头需 `Authorization: Bearer <JWT>`。角色 `statistics`（统计部）与 `warehouse`（库房）权限不同，见各接口说明。\n\n**Swagger**：展开接口后可用 **Try it out** 调试；**Curl** 区会按当前参数生成命令；**Example Value** 为示例请求/响应体。",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/", "description": "当前服务（与浏览器同源）" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 token"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录或 Token 无效",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" },
            "example": { "error": "未登录或缺少 Token" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/warehouses": {
      "get": {
        "tags": ["库房"],
        "summary": "库房列表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "按代码/名称/地址模糊搜索"
          }
        ],
        "responses": {
          "200": {
            "description": "库房列表",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "code": "DEF-001",
                    "name": "默认库房",
                    "address": "",
                    "createdAt": "2026-04-24 08:00:00",
                    "updatedAt": "2026-04-24 08:00:00"
                  }
                ]
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["库房"],
        "summary": "创建库房",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "example": { "code": "WH-002", "name": "二号库", "address": "某市某区" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "example": {
                  "id": 2,
                  "code": "WH-002",
                  "name": "二号库",
                  "address": "某市某区",
                  "createdAt": "2026-04-24 10:00:00",
                  "updatedAt": "2026-04-24 10:00:00"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/admin/warehouses/{id}": {
      "get": {
        "tags": ["库房"],
        "summary": "库房详情",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "单条库房",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "code": "DEF-001",
                  "name": "默认库房",
                  "address": "",
                  "createdAt": "2026-04-24 08:00:00",
                  "updatedAt": "2026-04-24 08:00:00"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "不存在",
            "content": {
              "application/json": { "example": { "error": "库房不存在" } }
            }
          }
        }
      },
      "put": {
        "tags": ["库房"],
        "summary": "更新库房",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": {
            "application/json": { "example": { "name": "默认库房（改名）", "address": "新地址" } }
          }
        },
        "responses": {
          "200": {
            "description": "更新后",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "code": "DEF-001",
                  "name": "默认库房（改名）",
                  "address": "新地址",
                  "createdAt": "2026-04-24 08:00:00",
                  "updatedAt": "2026-04-24 11:00:00"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["库房"],
        "summary": "删除库房",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "已删除" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}